스토리북 설정에 전역 스타일 적용
