--- .storybook/preview.ts
@@ -1,17 +1,0 @@
-// Replace your-framework with the framework you are using (e.g., react, vue3)
-import { Preview } from "@storybook/react";
-
-const preview: Preview = {
-  parameters: {
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/,
-      },
-    },
-  },
-
-  tags: ["autodocs"]
-};
-
-export default preview;
